refactor(parser): move checkParserErrors to helper file
